Renderer is now produced via generated code
This greatly reduces the amount of code to keep track of, and lets me
specify the rendering process in great detail. It is a first step toward
something resembling programmable shaders (but compiled in rather than
built at runtime). It also makes me much more confident that there are
no undiscovered typos in the rendering engine.